REFACTOR: pokemon store
