+++ src/lib/store.js
@@ -4,16 +4,8 @@ import {
   getPokemons as getPokemonsApi,
   getAllPokemons as getAllPokemonsApi,
   getDataForPokemon as getDataForPokemonApi,
-  getSpeciesData as getSpeciesDataApi,
 } from '@/api/pokemon';
-import { getPokemonEvolutions as getPokemonEvolutionsApi } from '@/api/evolutions';
-import { getPokemonColorTranslation as getPokemonColorTranslationApi } from '@/api/colors';
-import { getPokemonShapeTranslation as getPokemonShapeTranslationApi } from '@/api/shapes';
-import { getPokemonHabitatTranslation as getPokemonHabitatTranslationApi } from '@/api/habitat';
-import { getPokemonStatTranslation as getPokemonStatTranslationApi } from '@/api/stats';
-import { getPokemonTypeTranslation as getPokemonTypeTranslationApi } from '@/api/types';
-import { getAllCharacteristics } from '@/store/mutations/characteristics';
-import { getPokemonVariants, setVariant } from '@/store/mutations/variations';
+import { setVariant } from '@/store/mutations/variations';
 
 const state = Vue.observable({
   allPokemons: [],
@@ -23,7 +15,6 @@ const state = Vue.observable({
     pokemons: [],
     nextUrl: '',
   },
-  pokemon: new Map(),
 });
 
 export default {
@@ -112,89 +103,6 @@ export default {
 
   async pokemonIsVariant(name) {
     return name.includes('-') && !!(await getPokemonApi(name.split('-')[0]));
-  },
-
-  async getPokemon(pokemonId) {
-    if (state.pokemon?.has(pokemonId)) {
-      return;
-    }
-
-    const pokemon = await getPokemonApi(pokemonId);
-
-    const evolutions = await getPokemonEvolutionsApi(pokemonId);
-    const { flavorTexts, color, shape, generation, habitat } = pokemon.species
-      .url
-      ? await getSpeciesDataApi(pokemon.species.url)
-      : [];
-    let highestStatName = '';
-    let highestStatValue = 0;
-    const stats = await Promise.all(
-      pokemon.stats.map(async (s) => {
-        if (s.base_stat > highestStatValue) {
-          highestStatName = s.stat.name;
-          highestStatValue = s.base_stat;
-        }
-        return {
-          name: await getPokemonStatTranslationApi(s.stat.name),
-          value: s.base_stat,
-        };
-      })
-    );
-    let characteristic = '';
-    (getAllCharacteristics().get(highestStatName) ?? []).map((c) => {
-      if (c.possibleValues.includes(Math.floor(highestStatValue / 5))) {
-        characteristic = c.description;
-      }
-    });
-    const variants = [];
-    await Promise.all(
-      getPokemonVariants(pokemonId)?.map(async (pokemon) => {
-        const { name, sprites } = await getPokemonApi(pokemon);
-        if (sprites.front_default) {
-          variants.push({
-            name: name.replace(`${pokemonId}-`, '').replace('-', ' '),
-            image: sprites.front_default,
-          });
-        }
-      }) ?? []
-    );
-    const image = pokemon.sprites.other.dream_world.front_default;
-    const smallImage = pokemon.sprites.front_default;
-    const types = await Promise.all(
-      pokemon.types.map(async (t) => ({
-        name: t.type.name,
-        translated: await getPokemonTypeTranslationApi(t.type.name),
-      }))
-    );
-    const name = pokemon.name;
-    const id = pokemon.id;
-    const height = pokemon.height;
-    const weight = pokemon.weight;
-
-    state.pokemon.set(name, {
-      id,
-      name,
-      image,
-      smallImage,
-      stats,
-      types,
-      evolutions,
-      flavorTexts,
-      characteristic,
-      height,
-      weight,
-      color: {
-        name: color,
-        translated: await getPokemonColorTranslationApi(color),
-      },
-      shape: await getPokemonShapeTranslationApi(shape),
-      generation,
-      habitat: {
-        name: habitat,
-        translated: await getPokemonHabitatTranslationApi(habitat),
-      },
-      variants,
-    });
   },
 
   getAllPokemonsReplace() {
